Extract `cio_markdown` subpackage
This allows us to isolate our markdown rendering code in a dedicated package, which could potentially have a positive effect on the compile times of the full project. It also means we can iterate on this part of the code independent from the other parts, and with a well-specified API interface.
